@@ -50,13 +50,14 @@
     "thorax": "^2.0.3",
     "underscore": "^1.6.0",
     "victory": "^0.20.0",
-    "webpack-stream": "^3.2.0"
+    "webpack-stream": "^6.0.0"
   },
   "devDependencies": {
-    "babel-core": "^5.8.25",
-    "babel-eslint": "^4.1.3",
-    "babel-loader": "^5.3.2",
-    "babel-plugin-react-transform": "^1.1.1",
+    "@babel/core": "^7.0.0",
+    "@babel/eslint-parser": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "babel-loader": "^8.0.0",
+    "bluebird": "^2.9.33",
     "bower": "^1.8.8",
     "gulp-browserify": "~0.4.2",
     "gulp-clean": "~0.2.4",
@@ -85,7 +86,8 @@
     "react-transform-catch-errors": "^1.0.0",
     "react-transform-hmr": "^1.0.1",
     "redbox-react": "1.1.1",
-    "webpack": "^1.12.2",
+    "webpack": "^5.0.0",
+    "webpack-cli": "^4.3.1",
     "webpack-dev-middleware": "^1.2.0",
     "webpack-hot-middleware": "^2.4.1"
   },
