@@ -62,15 +62,15 @@ function fbLoginPrompt() {
   FB.getLoginStatus(function(response) {
     if (response.status !== 'connected') {
       return FB.login(function(response) {
-      if (response.authResponse) {
-        return facebookLoginOkHandler(response).then(dfd.resolve, dfd.reject);
-      } else {
-        return dfd.reject();
-      }
-    }, {
-      return_scopes: true, // response should contain the scopes the user allowed
-      scope: ['public_profile', 'user_location', 'user_friends','email'].join(',')
-    });
+        if (response.authResponse) {
+          return facebookLoginOkHandler(response).then(dfd.resolve, dfd.reject);
+        } else {
+          return dfd.reject();
+        }
+      }, {
+        return_scopes: true, // response should contain the scopes the user allowed
+        scope: ['public_profile', 'user_location', 'user_friends','email'].join(',')
+      });
     } else {
       if (response.authResponse) {
         return facebookLoginOkHandler(response).then(dfd.resolve, dfd.reject);
